requirements.txt
streamlit
openai-whisper
pydub
jiwer
torch
torchaudio
ffmpeg-python
